Remove DuplicateForm #6199
